requirements.txt
Django==1.7.1
django-gravatar2==1.1.4
django-tastypie==0.12.1
Markdown==2.5.1
pytest==2.6.4
pytest-django==2.7.0
